+++ apps/cold-graphql/src/backend/schema/ecoinvent-import.ts
@@ -1,0 +1,46 @@
+import { Entity, Field, ID, RelationshipField, graphweaverMetadata } from '@exogee/graphweaver';
+import { ISODateStringScalar } from '@exogee/graphweaver-scalars';
+import { MikroBackendProvider } from '@exogee/graphweaver-mikroorm';
+import { EcoinventDatum } from './ecoinvent-datum';
+import { EcoinventImportsProcessingStatus, EcoinventImport as OrmEcoinventImport } from '../entities';
+import { connection } from '../database';
+
+graphweaverMetadata.collectEnumInformation({ target: EcoinventImportsProcessingStatus, name: 'EcoinventImportsProcessingStatus' });
+
+@Entity<EcoinventImport>('EcoinventImport', {
+	provider: new MikroBackendProvider(OrmEcoinventImport, connection),
+})
+export class EcoinventImport {
+	@Field(() => ID, { primaryKeyField: true })
+	id!: string;
+
+	@Field(() => String)
+	bucket!: string;
+
+	@Field(() => String)
+	key!: string;
+
+	@Field(() => String)
+	activityName!: string;
+
+	@Field(() => String)
+	location!: string;
+
+	@Field(() => String)
+	referenceProduct!: string;
+
+	@Field(() => String, { nullable: true })
+	jobStatus?: string;
+
+	@Field(() => EcoinventImportsProcessingStatus)
+	processingStatus: string = EcoinventImportsProcessingStatus.PENDING;
+
+	@Field(() => ISODateStringScalar)
+	createdAt!: Date;
+
+	@Field(() => ISODateStringScalar, { nullable: true })
+	updatedAt?: Date;
+
+	@RelationshipField<EcoinventDatum>(() => [EcoinventDatum], { relatedField: 'ecoinventImport' })
+	ecoinventData!: EcoinventDatum[];
+}