+++ libs/nest/prisma/migrations/20250225023343_add_pending_procesing_status/migration.sql
@@ -1,0 +1,2 @@
+-- AlterEnum
+ALTER TYPE "processing_status" ADD VALUE 'PENDING';